Give credit where credit is due
When I was about to quit my University Assistant Professor position, Lucas showed me a project he'd like to work on for a final assignment. It was based on Scapy. Later on I flashed my router with OpenWRT and found that I could give some use to Scapy and started this project.